requirements.txt


# HTTP requests
requests

# HTML parsing and cleaning
beautifulsoup4
lxml

# Markdown conversion
markdownify

# Environment variable management
python-dotenv
croniter

# Logging (standard library, but can add rich for better logs)
rich

# OpenAI API
openai

# Scheduling (for local testing, not needed in production on App Platform)
schedule

# Testing
pytest
